Changed strcucture LinkedList to List and everything is working fine

+++ EstudoMusicaCs/Musica.cs
@@ -5,10 +5,10 @@ namespace EstudoMusica{
     public class Musica{
         public string Nome { get; private set; }
         public string Genero { get; private set; }
-        public float Duracao { get; private set; }
+        public int Duracao { get; private set; }
         public string DataLancamento{ get; private set; }
 
-        public Musica(string nome, string genero, float duracao, string lancamento){
+        public Musica(string nome, string genero, int duracao, string lancamento){
             this.Nome = nome;
             this.Genero = genero;
             this.Duracao = duracao;
